+++ Replica/replica.c
@@ -925,7 +925,7 @@ int isTagExist(GHashTable * funmetatable , char *filename , TAG* secureTag )
 
 void *bind_thread(void *port)
 {
-    int PORT=(int)port;
+    int PORT=(intptr_t)port;
 
     //Socket descriptor for the Replica
     int         sock;
@@ -952,7 +952,7 @@ void *bind_thread(void *port)
     /*Initialize socket structure */
     serv_addr.sin_family=AF_INET;
     serv_addr.sin_addr.s_addr=htonl(INADDR_ANY);
-    serv_addr.sin_port=htons(port);
+    serv_addr.sin_port=htons((intptr_t)port);
 
     //Point to struct serv_addr.
     servPtr=(struct sockaddr *) &serv_addr;
@@ -1001,7 +1001,7 @@ void *bind_thread(void *port)
         }
 
 
-        if(err=pthread_create(&tid , NULL , &accept_thread ,  (void *) newsock))
+        if(err=pthread_create(&tid , NULL , &accept_thread ,  (void *)(intptr_t) newsock))
         {
             perror2("pthread_create for accept_thread" , err);
             exit(1);
@@ -1017,7 +1017,7 @@ void *accept_thread(void *accept_sock)
     //socket descriptor for each client
     int acptsock;
     //Retrieve the socket that passed from pthread_create
-    acptsock= ((int)(accept_sock));
+    acptsock= (intptr_t)(accept_sock);
 
     int         bytes;
     int         len;
@@ -1353,7 +1353,7 @@ int main(int argc , char  *argv[])
     signal(SIGSEGV, signal_handler);
 
     //Create a thread for the bind.
-    if(err=pthread_create(&tid , NULL ,(void *) &bind_thread , (void *)port))
+    if(err=pthread_create(&tid , NULL ,(void *) &bind_thread , (void *)(intptr_t)port))
     {
         perror2("pthread_create", err);
         exit(1);
